+++ app/hooks/useGetPosts.ts
@@ -22,11 +22,10 @@ function useGetPosts() {
       setIsLoading(false);
     } else {
       axios
-        .get(hostName + url)
+        .get(`${hostName}${url}`)
         .then((res) => {
           const data = res.data;
           if (Array.isArray(data)) {
-            console.log(data);
             setPosts(data);
             setIsLoading(false);
           } else {
